+++ packages/headless/src/case-assist.index.ts
@@ -26,7 +26,10 @@ export * from './features/document-suggestion/document-suggestion-actions-loader
 export * from './features/case-assist/case-assist-analytics-actions-loader';
 
 // Controllers
-export type {Controller} from './controllers/controller/headless-controller';
+export type {
+  Controller,
+  Subscribable,
+} from './controllers/controller/headless-controller';
 export {buildController} from './controllers/controller/headless-controller';
 
 export type {