+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.6](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@0.7.5...@coveo/atomic-react@0.7.6) (2022-03-28)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [0.7.5](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@0.7.4...@coveo/atomic-react@0.7.5) (2022-03-25)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [0.7.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@0.7.3...@coveo/atomic-react@0.7.4) (2022-03-24)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [0.7.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@0.7.2...@coveo/atomic-react@0.7.3) (2022-03-24)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [0.7.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@0.7.1...@coveo/atomic-react@0.7.2) (2022-03-22)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
 ## [0.7.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@0.7.0...@coveo/atomic-react@0.7.1) (2022-03-18)
 
 **Note:** Version bump only for package @coveo/atomic-react